% Generated by roxygen2 (4.0.1): do not edit by hand
\name{find_place}
\alias{find_place}
\title{Geolocate an address}
\usage{
find_place(location, name, key = getOption("RYDN_KEY"),
  secret = getOption("RYDN_SECRET"))
}
\arguments{
\item{location}{The location name or address you want to lookup}

\item{name}{The name of a point of interest to lookup}

\item{key}{The Yahoo! Developer Network BOSS application key. Register
for an API key at \link{https://developer.yahoo.com/boss/}}

\item{secret}{The Yahoo! Developer Network BOSS application secret}
}
\description{
Lookup the latitude and longitude of a location using the place's
name or address.
}
\details{
There are charges associated with the use of this function set by
your Yahoo! BOSS application. See
\link{https://info.yahoo.com/legal/us/yahoo/boss/pricing/} for pricing
details.
}

